chore(grdb): retire CloudKit-side analysis compute helpers
The four §3.4 analysis methods are now SQL-aggregated on GRDB and the
production AnalysisRepository is GRDBAnalysisRepository. Drop the
CloudKitAnalysisRepository compute helpers, port the small Swift-only
helpers they still served (best-fit regression, scheduled-transaction
extrapolation, per-instance leg conversion) into sibling extensions on
the GRDB repository, and move InvestmentValueSnapshot alongside its
GRDB consumer. Also tighten doc comments and tests that name the
deleted CloudKit helpers.
